requirements.txt
black==21.8b0
click==8.0.1
flake8==3.9.2
mccabe==0.6.1
mypy==0.910
mypy-extensions==0.4.3
numpy==1.21.2
pandas==1.3.2
pathspec==0.9.0
platformdirs==2.3.0
pycodestyle==2.7.0
pyflakes==2.3.1
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
regex==2021.8.28
six==1.16.0
toml==0.10.2
tomli==1.2.1
types-PyYAML==5.4.10
typing-extensions==3.10.0.2
